Refactored Commands and Queries into shared module
