+++ src/api-authentication/src/test/java/org/rbernalop/apiauthentication/securityuser/application/finder/SecurityUserFinderTest.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.rbernalop.apiauthentication.user.domain.aggregate.UserMother;
-import org.rbernalop.apiauthentication.shared.application.userfinder.username.FindUserByUsernameQuery;
-import org.rbernalop.apiauthentication.user.application.finder.username.FindUserByUsernameQueryMother;
-import org.rbernalop.apiauthentication.shared.application.userfinder.username.FindUserByUsernameResponse;
-import org.rbernalop.apiauthentication.user.application.finder.username.FindUserByUsernameResponseMother;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameQuery;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameQueryMother;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameResponse;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameResponseMother;
 import org.rbernalop.apiauthentication.user.domain.aggregate.User;
 import org.rbernalop.shared.domain.bus.query.QueryBus;
 import org.rbernalop.shared.infrastructure.testing.UnitTestCase;
