+++ src/api-authentication/src/test/java/org/rbernalop/apiauthentication/user/application/finder/username/FindUserByUsernameQueryHandlerTest.java
@@ -3,8 +3,10 @@ package org.rbernalop.apiauthentication.user.application.finder.username;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
-import org.rbernalop.apiauthentication.shared.application.userfinder.username.FindUserByUsernameQuery;
-import org.rbernalop.apiauthentication.shared.application.userfinder.username.FindUserByUsernameResponse;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameQuery;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameQueryMother;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameResponse;
+import org.rbernalop.apiauthentication.shared.application.user.finder.username.FindUserByUsernameResponseMother;
 import org.rbernalop.apiauthentication.user.domain.aggregate.UserMother;
 import org.rbernalop.apiauthentication.user.domain.aggregate.User;
 import org.rbernalop.apiauthentication.user.domain.port.UserRepository;
